+++ src/migrations/Migrators/BlogFieldGroupMigrator.php
@@ -5,6 +5,8 @@ namespace matfish\Blogify\migrations\Migrators;
 
 use Craft;
 use craft\models\FieldGroup;
+use craft\records\FieldGroup as FieldGroupRecord;
+use matfish\Blogify\Handles;
 
 class BlogFieldGroupMigrator extends Migrator
 {
@@ -13,7 +15,7 @@ class BlogFieldGroupMigrator extends Migrator
         blogify_log("Adding Blog Fields group");
 
         $group = new FieldGroup([
-            'name' => 'Blog Fields'
+            'name' => Handles::BLOG_FIELDS_GROUP_NAME
         ]);
 
         return Craft::$app->fields->saveGroup($group);
@@ -22,7 +24,19 @@ class BlogFieldGroupMigrator extends Migrator
 
     public static function remove(): bool
     {
-        // removed by BlogChannelMigrator?
-        return true;
+        $name = Handles::BLOG_FIELDS_GROUP_NAME;
+
+        $group = FieldGroupRecord::findOne([
+            'name' => $name
+        ]);
+
+        if ($group) {
+            blogify_log("Removing {$name} group");
+            $group->delete();
+            return true;
+        }
+
+        blogify_log("Field group {$name} not found. Skipping.");
+        return false;
     }
 }